fix(new-client): added browser unsupported message for browsers without BigInt
