+++ src/new-client/index.html
@@ -16,6 +16,7 @@
     <link rel="apple-touch-icon" href="/touch-icon-iphone.png" />
     <link rel="manifest" href="/manifest.json" />
     <title>Reactive Trader®</title>
+    <%- injectScript %>
   </head>
 
   <body>
