+++ src/new-client/package.json
@@ -130,6 +130,7 @@
     "tslib": "^1.14.1",
     "typescript": "^4.1.2",
     "vite": "^2.1.5",
+    "vite-plugin-html": "^2.1.2",
     "wait-on": "^5.3.0"
   },
   "prettier": {
